fixing some bugs with control and anaylsing calibration data in matlab

+++ lib/controller/gainScheduleData.h
@@ -7,9 +7,9 @@
 
 struct gainScheduleData
 {
-    uint16_t height;                  // Number of rows read from the file
-    static const uint8_t width = 4;   // Number of columns (fixed at 4)
-    float data[MAX_GAIN_ROWS][width]; // 2D array to store the data
+    uint8_t height;                    // Number of rows read from the file
+    static const uint8_t width = 4;    // Number of columns (fixed at 4)
+    double data[MAX_GAIN_ROWS][width]; // array to store the data. [P, I, D, pressure]
 };
 
 #endif // GAIN_SCHEDULE_DATA_H
